Change api to use index
