+++ packages/now-build-utils/.gitignore
@@ -1,4 +1,1 @@
-# ignore generated files
-*.js
-*.d.ts
-fs/**
+dist
