+++ packages/now-build-utils/.npmignore
@@ -1,2 +1,3 @@
+/src
 /test
 tmp
