+++ packages/now-build-utils/src/types.ts
@@ -1,0 +1,9 @@
+export interface File {
+  type: string;
+  mode: number;
+  toStream: () => NodeJS.ReadableStream;
+}
+
+export interface Files {
+  [filePath: string]: File
+}
